Fixing driver NPEs in tear down methods for JUnit
At this moment, every time a JUnit test fails and the driver is left as
null, an extra NPE will be thrown each time driver.quit() is called. This
commit fixes this issue by asking every time if driver is different from
null before quitting the driver.
Also, there is a solution for issue #56.